Change CustomOllama class to match langchain update
